requirements.txt
viktor==14.20.0
pydantic
plotly
dedent
kaleido==0.2.1; platform_system == "Linux"
kaleido==0.1.0.post1; platform_system == "Windows"
python-dotenv
pandas
openpyxl
openai
instructor